+++ packages/portal-components/src/index.ts
@@ -1,0 +1,15 @@
+export { ComponentExample } from "./ComponentExample";
+export { DevExample } from "./DevExample";
+
+export interface Dictionary<T> {
+    [key: string]: T;
+}
+export interface ExampleComponentProps {
+    boolValues?: Dictionary<boolean>;
+    choiceValues?: Dictionary<string>;
+}
+export interface ChoiceProp {
+    name: string;
+    values: Array<string>;
+    defaultValue: number;
+}