+++ packages/react-hooks/documentation/MutationObserverExample.tsx
@@ -17,7 +17,7 @@ const MutationObserverExample = () => {
 
     useMutationObserver(mutationTargetRef, onObservation, { characterData: true, subtree: true });
     return (
-        <section className="hooks-example jkl-spacing--bottom-3">
+        <section>
             <button
                 className="jkl-button jkl-button--secondary jkl-spacing--right-1"
                 onClick={() => setIsOpen(!isOpen)}
@@ -27,12 +27,13 @@ const MutationObserverExample = () => {
             <button className="jkl-button jkl-button--tertiary" onClick={() => appendToMutationList([])}>
                 Nullstill liste
             </button>
-            <p ref={mutationTargetRef} className="jkl-spacing--top-1 hooks-example__text">
+            <p ref={mutationTargetRef} className="jkl-spacing--top-1 jkl-body">
                 Nåværende status: <strong>{isOpen ? "Åpen" : "Lukket"}</strong>
+                <br />
+                Liste over endringer:
             </p>
-            <p className="hooks-example__text">Liste over endringer:</p>
             {listOfMutations.length !== 0 && (
-                <ul className="jkl-list hooks-example__list">
+                <ul className="jkl-list jkl-list--unordered jkl-body">
                     {listOfMutations.map((item, idx) => (
                         <li className="jkl-list__item" key={idx}>
                             {item}