+++ packages/react-hooks/documentation/index.tsx
@@ -1,10 +1,40 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import Example from "./Example";
-import "../../webfonts/documentation/internal.scss";
+
+// Import core styles, webfonts and tab listener (same for all components):
 import { initTabListener } from "@fremtind/jkl-core";
+import "@fremtind/jkl-core/core.scss";
+import "../../webfonts/documentation/internal.scss";
+
+// Imports required for showing example (same for all components):
+import { DevExample } from "@fremtind/jkl-portal-components";
+import "@fremtind/jkl-portal-components/dev-example.min.css";
+import "@fremtind/jkl-radio-button/radio-button.min.css";
+import "@fremtind/jkl-checkbox/checkbox.min.css";
+import "@fremtind/jkl-button/button.min.css";
+import "@fremtind/jkl-list/list.min.css";
+import "@fremtind/jkl-card/card.min.css";
+
+// Import actual example and component stylesheet (specific for this component):
+import AnimationExample from "./AnimationExample";
+import ClickOutsideExample from "./ClickOutsideExample";
+import FocusOutsideExample from "./FocusOutsideExample";
+import IntersectionObserverExample from "./IntersectionObserverExample";
+import KeyListenerExample from "./KeyListenerExample";
+import MutationObserverExample from "./MutationObserverExample";
+import "@fremtind/jkl-datepicker/datepicker.css";
 
 initTabListener();
 
 const mountNode = document.getElementById("app");
-ReactDOM.render(<Example />, mountNode);
+ReactDOM.render(
+    <>
+        <DevExample component={AnimationExample} />
+        <DevExample component={ClickOutsideExample} />
+        <DevExample component={FocusOutsideExample} />
+        <DevExample component={KeyListenerExample} />
+        <DevExample component={MutationObserverExample} />
+        <DevExample component={IntersectionObserverExample} />
+    </>,
+    mountNode,
+);