+++ portal/src/components/Typography/FormatProvider.tsx
@@ -1,0 +1,42 @@
+import React, { ReactNode } from "react";
+import { MDXProvider } from "@mdx-js/react";
+import { Link } from "@fremtind/jkl-typography-react";
+import { OrderedList, UnorderedList, ListItem } from "@fremtind/jkl-list-react";
+import { ComponentExample } from "@fremtind/jkl-portal-components";
+import "@fremtind/jkl-portal-components/component-example.scss";
+import { DoDontExample } from "../DoDontExample";
+import "./typography.scss";
+
+import {
+    PageTitle,
+    HeadingLarge,
+    HeadingMedium,
+    HeadingSmall,
+    HeadingXS,
+    Paragraph,
+    ArticleLead as Ingress,
+} from "../Typography";
+
+const components = {
+    h1: PageTitle,
+    h2: HeadingLarge,
+    h3: HeadingMedium,
+    h4: HeadingSmall,
+    h5: HeadingXS,
+    p: Paragraph,
+    ul: UnorderedList,
+    ol: OrderedList,
+    li: ListItem,
+    a: Link,
+    Ingress,
+    ComponentExample,
+    DoDontExample,
+};
+
+interface Props {
+    children: ReactNode;
+}
+
+export function FormatProvider({ children }: Props) {
+    return <MDXProvider components={components}>{children}</MDXProvider>;
+}