+++ packages/datepicker-react/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.1](https://github.com/fremtind/jokul/compare/@fremtind/jkl-datepicker-react@1.4.0...@fremtind/jkl-datepicker-react@1.4.1) (2020-03-06)
+
+**Note:** Version bump only for package @fremtind/jkl-datepicker-react
+
+
+
+
+
+# [1.4.0](https://github.com/fremtind/jokul/compare/@fremtind/jkl-datepicker-react@1.3.1...@fremtind/jkl-datepicker-react@1.4.0) (2020-03-06)
+
+
+### Features
+
+* **portal:** implement dev and portal examples ([857f2e2](https://github.com/fremtind/jokul/commit/857f2e29359f9442a7da0c977a94d63dfbdfc194))
+
+
+
+
+
+## [1.3.1](https://github.com/fremtind/jokul/compare/@fremtind/jkl-datepicker-react@1.3.0...@fremtind/jkl-datepicker-react@1.3.1) (2020-03-05)
+
+**Note:** Version bump only for package @fremtind/jkl-datepicker-react
+
+
+
+
+
 # [1.3.0](https://github.com/fremtind/jokul/compare/@fremtind/jkl-datepicker-react@1.2.3...@fremtind/jkl-datepicker-react@1.3.0) (2020-02-24)