+++ packages/text-input/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.3](https://github.com/fremtind/jokul/compare/@fremtind/jkl-text-input@1.3.2...@fremtind/jkl-text-input@1.3.3) (2020-03-06)
+
+**Note:** Version bump only for package @fremtind/jkl-text-input
+
+
+
+
+
+## [1.3.2](https://github.com/fremtind/jokul/compare/@fremtind/jkl-text-input@1.3.1...@fremtind/jkl-text-input@1.3.2) (2020-03-06)
+
+**Note:** Version bump only for package @fremtind/jkl-text-input
+
+
+
+
+
+## [1.3.1](https://github.com/fremtind/jokul/compare/@fremtind/jkl-text-input@1.3.0...@fremtind/jkl-text-input@1.3.1) (2020-03-05)
+
+**Note:** Version bump only for package @fremtind/jkl-text-input
+
+
+
+
+
 # [1.3.0](https://github.com/fremtind/jokul/compare/@fremtind/jkl-text-input@1.2.2...@fremtind/jkl-text-input@1.3.0) (2020-02-19)