+++ packages/toggle-switch-react/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.1](https://github.com/fremtind/jokul/compare/@fremtind/jkl-toggle-switch-react@2.1.0...@fremtind/jkl-toggle-switch-react@2.1.1) (2020-03-06)
+
+**Note:** Version bump only for package @fremtind/jkl-toggle-switch-react
+
+
+
+
+
+# [2.1.0](https://github.com/fremtind/jokul/compare/@fremtind/jkl-toggle-switch-react@2.0.2...@fremtind/jkl-toggle-switch-react@2.1.0) (2020-03-06)
+
+
+### Features
+
+* **portal:** implement dev and portal examples ([857f2e2](https://github.com/fremtind/jokul/commit/857f2e29359f9442a7da0c977a94d63dfbdfc194))
+
+
+
+
+
+## [2.0.2](https://github.com/fremtind/jokul/compare/@fremtind/jkl-toggle-switch-react@2.0.1...@fremtind/jkl-toggle-switch-react@2.0.2) (2020-03-05)
+
+**Note:** Version bump only for package @fremtind/jkl-toggle-switch-react
+
+
+
+
+
 ## [2.0.1](https://github.com/fremtind/jokul/compare/@fremtind/jkl-toggle-switch-react@2.0.0...@fremtind/jkl-toggle-switch-react@2.0.1) (2020-02-19)
 
 **Note:** Version bump only for package @fremtind/jkl-toggle-switch-react